src/requirements.txt
dj-database-url==0.5.0
Django==2.1.5
django-crispy-forms==1.7.2
django-filter==2.0.0
django-webpack-loader==0.6.0
djangorestframework==3.9.0
gunicorn==19.9.0
Markdown==3.0.1
mysqlclient==1.3.14
Pillow==5.4.1
pytz==2018.9
